+++ ESLZ/SRV-Linux.tf
@@ -22,6 +22,6 @@ module "linux_VMs" {
   resource_groups   = local.resource_groups_all
   subnets           = local.subnets
   user_data         = try(each.value.user_data, false) != false ? base64encode(file("${path.cwd}/${each.value.user_data}")) : null
-  custom_data       = try(each.value.custom_data, false) != false ? each.value.custom_data == "install-ca-certs" ? each.value.custom_data : base64encode(file("${path.cwd}/${each.value.custom_data}")) : null
+  custom_data       = try(each.value.custom_data, false) != false ? contains(["install-ca-certs", "cloud-init-default"], each.value.custom_data) || can(regex("^https?://", each.value.custom_data)) ? each.value.custom_data : base64encode(file("${path.cwd}/${each.value.custom_data}")) : null
   tags              = var.tags
 }